backend/requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
openai-whisper==20231117
python-dotenv==1.0.0
requests==2.31.0 